requirements.txt
pymupdf
pytesseract
pillow
opencv-python-headless 
numpy
